% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_link.R
\docType{data}
\name{connectors_list}
\alias{connectors_list}
\title{list of proper names connectors}
\format{
An object of class \code{list} of length 6.
}
\usage{
connectors_list
}
\description{
list of connectors between proper names in different languages
}
\keyword{datasets}
