% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccur.R
\name{cooccur}
\alias{cooccur}
\title{Get cooccurrence of all words}
\usage{
cooccur(
  text,
  token_by = "sentence",
  sw = "",
  lower = TRUE,
  loop = FALSE,
  count = TRUE
)
}
\arguments{
\item{text}{Text}

\item{token_by}{Tokenize by sentence or paragraph}

\item{sw}{Stopwords to be removed}

\item{lower}{Convert words to lowercase. If the text is passed in all lowercase, it can return false sentence and paragraph tokenization.}

\item{loop}{if FALSE, self referential nodes (e.g. n1=x and also n2=x) will be excluded. Default FALSE.}

\item{count}{Return count of words (default TRUE)}
}
\description{
Get cooccurrence of all words
}
\examples{
text |> cooccur()
}
