% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccur_word.R
\name{cooccur_words}
\alias{cooccur_words}
\title{Get cooccurrence of all words}
\usage{
cooccur_words(
  text,
  sw = "",
  token_by = "sentence",
  lower = TRUE,
  loop = FALSE,
  output = "df",
  count = TRUE
)
}
\arguments{
\item{text}{Text}

\item{sw}{Stopwords to be removed}

\item{token_by}{Tokenize by sentence or paragraph}

\item{lower}{Convert words to lowercase. If the text is passed in all
lowercase, it can return false sentence and paragraph tokenization. It is
advised to use lowercase.}

\item{loop}{if FALSE, self referential nodes (e.g. n1=x and also n2=x) will
be excluded. Default FALSE.}

\item{output}{as 1) a single tibble/dataframe ("tlb", "df", "tibble",
"datafame"); 2) as a list of dataframes with cooccurrence per vector
element ("lst" or "list"); or 3) as raw list. This format is the most raw
output of this function; 4) "df2", tibble/dataframe with the doc numbers.}

\item{count}{Return count of words (default TRUE)}
}
\description{
Get cooccurrence of all words
}
\examples{
txt <- "Lorem Ipsum. The Ipsum John. Dolor est"
txt |> cooccur_words()
}
