% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{count_graph}
\alias{count_graph}
\title{Count the graph frequency of co-occurences/Co-mentioning of triplets}
\usage{
count_graph(graph, loop = FALSE)
}
\arguments{
\item{graph}{a graph of co-occurrence}

\item{loop}{if TRUE, include loops. Default=FALSE}
}
\description{
Count the graph frequency of co-occurences/Co-mentioning of triplets
}
\examples{
DF <- data.frame(n1 = c("Sample1", "Sample2", "Sample3"), n2 = c("A1", "B1", "A2"), l = c("Value1", "Value2", "Value3"))
DF |> count_graph()
}
