% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_graphs.R
\name{count_graphs}
\alias{count_graphs}
\title{Count the graph frequency of co-occurences/Co-mentioning of triplets}
\usage{
count_graphs(graph, loop = FALSE)
}
\arguments{
\item{graph}{a graph of co-occurrence}

\item{loop}{if TRUE, include loops. Default=FALSE}
}
\description{
Count the graph frequency of co-occurences/Co-mentioning of triplets
}
\examples{
DF <- data.frame(n1 = c("Sample1", "Sample1", "Sample2", "Sample2"), n2 = c("A1", "Sample1", "B1", "B1"))
DF |> count_graphs()
}
