% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_functions.R
\name{count_vec}
\alias{count_vec}
\title{count a vector of elements}
\usage{
count_vec(vec, sort_n = TRUE)
}
\description{
count a vector of elements, arrange it or not in frequency order, and returns a tibble
}
\examples{
vec <- s2v("a a b c a b a z z z c d e")
vec |> count_vec()
}
