% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\docType{data}
\name{df_names}
\alias{df_names}
\title{Names of dataframes, to be used inside functions in arg %in% df_names}
\format{
An object of class \code{character} of length 4.
}
\usage{
df_names
}
\description{
Names of dataframes, to be used inside functions in arg %in% df_names
}
\keyword{datasets}
