% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{entity_list_2_graph}
\alias{entity_list_2_graph}
\title{from a vector of entities (generated by extract_entities2) returns a tibble/dataframe with co-occurence pairs}
\usage{
entity_list_2_graph(entities)
}
\arguments{
\item{entities}{a vector of entities generated by extract_entities2}
}
\description{
from a vector of entities (generated by extract_entities2) returns a tibble/dataframe with co-occurence pairs
}
\keyword{internal}
