% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{entity_list_2_graph2}
\alias{entity_list_2_graph2}
\title{from a list of entities (generated by extract_entities2) returns a tibble/dataframe with co-occurence pairs}
\usage{
entity_list_2_graph2(entities_list, count = TRUE)
}
\arguments{
\item{entities_list}{a list object of entities generated by extract_entities2}

\item{count}{if TRUE returns a count of co-occurences}
}
\description{
from a list of entities (generated by extract_entities2) returns a tibble/dataframe with co-occurence pairs
}
