% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_prince.R
\docType{data}
\name{ex_prince}
\alias{ex_prince}
\title{The Prince by Niccolò Machiavelli}
\format{
a data frame with The Prince book from Gutenberg Project separated
by chapters
}
\source{
<https://www.gutenberg.org/ebooks/1232>
}
\usage{
ex_prince
}
\description{
The Prince by Niccolò Machiavelli
}
\examples{

data(prince)
str(prince)

}
\keyword{datasets}
