% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{extract_entities2}
\alias{extract_entities2}
\title{extract_entity for each sentence}
\usage{
extract_entities2(pos_df)
}
\value{
a list of all entities in each sentence
}
\description{
extract_entity for each sentence
}
