% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{extract_entities_l}
\alias{extract_entities_l}
\title{Extract entities in list format, by sentence}
\usage{
extract_entities_l(DF, rgx_entities = "^(PER(SON)?|ORG|GPE)_")
}
\arguments{
\item{DF}{a dataframe generated by spacyr::spacy_parse}
}
\value{
a vector
}
\description{
extract entities in list, separated by sentence.
}
\examples{
t |> extract_entities()

}
