% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{extract_entities_v}
\alias{extract_entities_v}
\title{extract entities, returns a vector}
\usage{
extract_entities_v(DF, rgx_entities = "^(PER(SON)?|ORG|GPE)_")
}
\arguments{
\item{DF}{a dataframe generated by spacyr::spacy_parse}
}
\value{
a vector
}
\description{
extract entities from Spacy entities.
}
\examples{
t |> extract_entities_v()

}
