% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_entity_rb.R
\name{extract_entity_rb}
\alias{extract_entity_rb}
\title{A rule based entity extractor}
\usage{
extract_entity_rb(
  text,
  connect = connectors("misc"),
  sw = "the",
  underscore = TRUE
)
}
\arguments{
\item{text}{an input text}

\item{connect}{a vector of lowercase connectors. Use use your own, or use the function "connector" to obtain some patterns.}

\item{sw}{a vector of stopwords}

\item{underscore}{keep underscore to make compounded words a unique word?}
}
\description{
extracts the entity from a text using regex. This regex captures all uppercase words, words that begin with upper case. If there is sequence of this patterns together, this function also captures.
In the case of proper names with common lower case connectors like "Wwwww of Wwwww" this function also captures the connector and the subsequent uppercase words.
}
\examples{
"John Does lives in New York in United States of America." |> extract_entity()
"João Ninguém mora em São José do Rio Preto. Ele esteve antes em Sergipe" |> extract_entity(connect = connectors("pt"))
text |> extract_entity()
}
