% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{extract_graph_pos}
\alias{extract_graph_pos}
\title{extract graph of co-ocurrence from a POS dataframe}
\usage{
extract_graph_pos(pos_df, POS = c("NOUN", "PROPN"))
}
\arguments{
\item{pos_df}{a data.frame generated by spacyr::spacy_parse}

\item{POS}{the POS to be extracted. Defaul: NOUN and PROPN}
}
\description{
extract graph of co-ocurrence from a POS dataframe
}
