% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_graph_rb.R
\name{extract_graph_rb}
\alias{extract_graph_rb}
\title{rule based entity graph extractor}
\usage{
extract_graph_rb(
  text,
  using = "sentences",
  connect = connectors("misc"),
  sw = c("of", "the"),
  count = TRUE,
  loop = FALSE
)
}
\arguments{
\item{text}{an input text}

\item{using}{sentence or paragraph to tokenize}

\item{connect}{lowercase connectors, like the "von" in "John von Neumann".}

\item{sw}{stopwords vector.}

\item{count}{if TRUE (default) count the frequency of nodes and return it in the order of its frequency}

\item{loop}{if TRUE, it will not remove loops, a node pointing to itself.}
}
\description{
Extract a non directional graph based on co-occurrence in the token.
It extracts only if two entities are mentioned in the same token (sentence or paragraph)
}
\examples{
text <- "John Does lives in New York in United States of America. He  is a passionate jazz musician, often playing in local clubs."
extract_graph_rb(text)
}
