% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{extract_nodes}
\alias{extract_nodes}
\title{extract nodes from a dataframe of two columns of nodes'}
\usage{
extract_nodes(graph)
}
\description{
extract nodes from a dataframe of two columns of nodes'
}
\keyword{internal}
