% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{extract_triplets}
\alias{extract_triplets}
\title{from semgram output, join passive and active voices
# VER great..../GR_analise}
\usage{
extract_triplets(semgram)
}
\description{
from semgram output, join passive and active voices
# VER great..../GR_analise
}
