% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R, R/graph_operations.R
\name{filter_ego}
\alias{filter_ego}
\title{Filter a graph / create an ego graph}
\usage{
filter_ego(edges, nodes = NULL, filter_by, n_neighbours = 1)

filter_ego(edges, nodes = NULL, filter_by, n_neighbours = 1)
}
\arguments{
\item{edges}{an edge dataframe}

\item{nodes}{a node dataframe}

\item{filter_by}{a term to filter the ego graph}

\item{n_neighbours}{the number of neighbors}

\item{as_tbl}{if TRUE, return a tbl_graph, if FALSE, return an igraph object}
}
\description{
filter a graph / create an ego graph by term and by the number of its neighbors
}
\examples{
# creating sample data
nodes <- data.frame(id = 1:5, name = LETTERS[1:5])
edges <- data.frame(from = c(1, 1, 2, 3, 4, 1, 6, 7), to = c(2, 3, 4, 5, 5, 4, 7, 5))
filter_ego(edges, nodes, filter_by = 1, n_neighbours = 1)
filter_ego(edges, nodes, filter_by = 1, n_neighbours = 2)
# creating sample data
nodes <- data.frame(id = 1:5, name = LETTERS[1:5])
edges <- data.frame(from = c(1, 1, 2, 3, 4, 1, 6, 7), to = c(2, 3, 4, 5, 5, 4, 7, 5))
filter_ego(edges, nodes, filter_by = 1, n_neighbours = 1)
filter_ego(edges, nodes, filter_by = "jojo", n_neighbours = 2)
}
