% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_graph.R
\name{filter_graph}
\alias{filter_graph}
\title{Filter both nodes at the same time}
\usage{
filter_graph(DF, query, invert = FALSE)
}
\arguments{
\item{DF}{Dataframe with at least two columns}

\item{query}{the term to be filtered}

\item{invert}{if TRUE, invert the filter. Default FALSE.}
}
\description{
filter the first two columns of graph dataframe
}
\examples{
graph <- data.frame(n1 = c("A", "B", "C"), n2 = c("B", "C", "D"))
filter_graph(graph, "B")
filter_graph(graph, "B", invert = TRUE)
}
