% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_graph.R
\name{filter_graph_g}
\alias{filter_graph_g}
\title{Filter both nodes at same time using grep}
\usage{
filter_graph_g(DF, query, ic = T, invert = FALSE)
}
\arguments{
\item{DF}{Dataframe with at least two columns}

\item{query}{the term to be filtered}

\item{ic}{ignore case (default TRUE)}

\item{invert}{if TRUE, invert the filter. Default FALSE.}
}
\description{
filter the first two columns of graph dataframe
}
\examples{
graph <- data.frame(n1 = c("A", "B", "C"), n2 = c("B", "C", "D"))
filter_graph_g(graph, "b")
}
