% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{filter_nested_list}
\alias{filter_nested_list}
\title{function to delete from list all list elements with less than 2 elements}
\usage{
filter_nested_list(x, min_length = 2)
}
\description{
used inside cooccur_words(), to erase empty elements that do not form a pair.
}
