% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nodes.R
\name{filter_nodes}
\alias{filter_nodes}
\title{filter nodes in both columns, at the same time}
\usage{
filter_nodes(DF, query, invert = FALSE, ...)
}
\arguments{
\item{DF}{Dataframe with at least two columns}

\item{query}{String to filter the columns}

\item{invert}{if TRUE, invert the filter. Default FALSE.}

\item{...}{parameters for grepl}
}
\description{
Given a dataframe with at least two columns, filter the first two columns
at the same time. It uses grepl not equal comparision `==`, so it is
possible to use grepl parameters like ignore.case, etc.
}
\examples{
txt <- "Lorem Ipsum. The Ipsum John. Dolor est Lorem"
DF <- txt |> cooccur()
DF
DF |> filter_nodes()
}
