% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{filter_ppn}
\alias{filter_ppn}
\title{extract proper name and nouns from POS DF}
\usage{
filter_ppn(pos_df, POS = c("NOUN", "PROPN"))
}
\arguments{
\item{pos_df}{a data.frame generated by spacyr::spacy_parse}

\item{POS}{the POS to be extracted. Default: NOUN and PROPN}
}
\description{
extract proper name and nouns from POS DF
}
\examples{
pos <- "Mary Jane was the first to dance with Robert" |> spacy_parse()
pos |> filter_ppn()
}
