% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_nodes.R
\name{freq_nodes}
\alias{freq_nodes}
\title{extract the frequency of nodes in a text}
\usage{
freq_nodes(vert, text, lower = TRUE)
}
\arguments{
\item{vert}{a vector of nodes/terms}

\item{text}{an input text}

\item{lower}{Convert words to lowercase.}
}
\value{
a dataframe of nodes and its frequency
}
\description{
extract the frequency of nodes in a text
}
