% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{data}
\name{g}
\alias{g}
\title{A sample graph dataframe (tbl-graph)}
\format{
A tibble graph, from {tidygraphs} package
\describe{
  \item{nodes}{the nodes names}
  \item{edges}{the relationships between the nodes}
}
}
\usage{
g
}
\description{
A sample dataset used to illustrate the functionality of the package.
}
\examples{
data(g)
head(g)
}
\keyword{datasets}
