% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_link_type.R
\name{geom_link_type}
\alias{geom_link_type}
\title{use straight line or arc link between nodes}
\usage{
geom_link_type(
  edge_type = "arc",
  edge_width = 1,
  edge_color,
  edge_alpha,
  edge_cut,
  edge_bend = 0.5
)
}
\arguments{
\item{edge_type}{arc or line.}
}
\description{
use straight line or arc link between nodes
}
