% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_words_in_graph.R
\name{get_cooc}
\alias{get_cooc}
\title{get graph (co-ocurrence of entities and Nouns)}
\usage{
get_cooc(pos_df, pos_cat = c("NOUN", "ENTITY"))
}
\arguments{
\item{pos_df}{a POS dataframe}

\item{pos_cat}{the POS categories to be extracted. Default: NOUN and ENTITY}
}
\description{
from a POS dataframe (using `filter_by_query() |> parsePOS()` ) get the pairs of co-ocurrences.
}
\examples{
x <- txt_wiki |> filter_by_query("Police") 
x  <- x |>  parsePOS(only_entities=FALSE) 
get_cooc(x)
}
