% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_words_in_graph.R
\name{get_cooc_entities}
\alias{get_cooc_entities}
\title{Get graph (co-occurrence of entities)}
\usage{
get_cooc_entities(pos, loop = FALSE, freq = TRUE, lower_case = FALSE)
}
\arguments{
\item{pos}{a dataframe generated by filter_by_query() |> parsePOS()}

\item{loop}{if TRUE (default: FALSE), returns a loop (self reference) in the graph}

\item{freq}{if TRUE (default: TRUE), returns the count (frequency) of edges/co-occurences}
}
\description{
from a POS dataframe (using filter_by_query() |> parsePOS()) get the pairs of entities
}
\examples{
x <- txt_wiki |> filter_by_query("Police")
x <- x |> parsePOS()
get_cooc_entities(x)

# with loops /self-reference
get_cooc_entities(x, loop = TRUE)
get_cooc_entities(x, lower_case = TRUE)
}
