% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_entities}
\alias{get_entities}
\title{get entities from POS DF}
\usage{
get_entities(pos_df, as_list = TRUE, rgx_entities = "^(PER(SON)?|ORG|GPE)_")
}
\arguments{
\item{pos_df}{a pos data frame as generated by spacyr::spacy_parse()}
}
\description{
get entities from POS DF
}
\examples{
pos_txt <- sample_text |> spacyr::spacy_parse()
pos_txt  |> get_entities(as_list = FALSE)
pos_txt  |> get_entities()
}
