% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_words_in_graph.R
\name{get_graph_from_txt}
\alias{get_graph_from_txt}
\title{graph from text and query}
\usage{
get_graph_from_txt(text, query, i_c = TRUE, by_sentence = TRUE, loop = FALSE)
}
\arguments{
\item{text}{a vector of texts}

\item{query}{query to filter the text}

\item{by_sentence}{tokenize by sentence (default TRUE). If FALSE, tokenize by paragraph}

\item{loop}{if TRUE (default: FALSE), returns a loop (self reference) in the graph}

\item{ic}{ignore case (default TRUE)}
}
\description{
from text and a query, generates the graph dataframe of co-ocurrence frequency
}
\examples{
# loading data
data(package="txtnet")
# sample
txt_wiki[2:3]
txt_wiki |> get_graph_from_txt("York")
txt_wiki |> get_graph_from_txt("Police")
}
