% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operations.R
\name{get_neighbors}
\alias{get_neighbors}
\title{get the neighbors of a site}
\usage{
get_neighbors(graph, query, n = 1)
}
\description{
creates a sub graph with the node of reference and its Nth neighbors
}
\examples{
g
g |> plot()
get_neighbors(g, "Charlie", 1)
get_neighbors(g, "Charlie", 2) |> plot()
get_neighbors(g, "Alice", 1) |> plot()
get_neighbors(g, "Alice", 2) |> plot()
}
