% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operations.R
\name{get_neighbors_df}
\alias{get_neighbors_df}
\title{Get the neighbors of a word in a dataframe}
\usage{
get_neighbors_df(DF, query)
}
\description{
given a graph dataframe, find the neighbour nodes of the actual nodes, 
increasing the size of the graph
}
