% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operations.R
\name{get_node_id}
\alias{get_node_id}
\title{Search for the id of a node in a graph}
\usage{
get_node_id(graph, query)
}
\arguments{
\item{graph}{a tbl_graph}

\item{querie}{the name of the node}
}
\description{
Search for the id of a node in a graph
}
\examples{
# Building a graph dataframe
# Node data frame:
nodes <- data.frame(name = c("Alice", "Bob", "Charlie", "David", "John", "Mary"))
# Edge data frame
edges <- data.frame(from = c(1, 1, 2, 3, 2, 6), to = c(2, 3, 4, 4, 5, 5))
g <- tidygraph::tbl_graph(nodes = nodes, edges = edges)
get_node_id(g, "Bob")
}
