% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_words_in_graph.R
\name{get_pairs}
\alias{get_pairs}
\title{get graph (co-ocurrence of entities)}
\usage{
get_pairs(pos, loop = FALSE, freq = TRUE)
}
\arguments{
\item{pos}{a dataframe generated by filter_by_query() |> parsePOS()}

\item{loop}{if TRUE (default: FALSE), returns a loop (self reference) in the graph}

\item{freq}{if TRUE (default: TRUE), returns the count (frequency) of edges/co-occurences}
}
\description{
from a POS dataframe (using filter_by_query() |> parsePOS()) get the pairs of entities
}
\examples{
txt_wiki |> filter_by_query("Police") |>
  parsePOS() |>
  get_pairs()
txt_wiki |> filter_by_query("Police") |>
  parsePOS() |>
  get_pairs(loop = TRUE)
}
