% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_words_in_graph.R
\name{get_pairs2}
\alias{get_pairs2}
\title{get graph (co-ocurrence of entities and Nouns)}
\usage{
get_pairs2(pos_df, pos_cat = c("NOUN", "ENTITY"))
}
\arguments{
\item{pos_df}{a POS dataframe}

\item{pos_cat}{the POS categories to be extracted. Default: NOUN and ENTITY}
}
\description{
from a POS dataframe (using filter_by_query() |> parsePOS()) get the pairs of entities
}
\examples{
txt_wiki |> filter_by_query("Police") |>
  parsePOS(only_entities=FALSE) |>
  get_pairs2()
}
