% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_subs.R
\name{graph_subs}
\alias{graph_subs}
\title{Substitute node columns (columns 1 and 2) of a graph data frame with a
dictionary of substitutions.}
\usage{
graph_subs(DF, df_subs)
}
\arguments{
\item{DF}{a graph data frame}

\item{df_subs}{a data frame with the substitutions to be made, where the
first column is what to substitute and column 2 is what to substitute with.
It uses regular expressions.}
}
\description{
Substitute node columns (columns 1 and 2) of a graph data frame with a
dictionary of substitutions.
}
