% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{group_entity}
\alias{group_entity}
\title{Group a POS  of proper names}
\usage{
group_entity(DF)
}
\arguments{
\item{DF}{A data.frame generated by spacyr::spacy_parse. Crated by spacy_parse()}
}
\description{
Group a POS  of proper names
}
\examples{
# example in Portuguese language
spacy_initialize(model = "pt_core_news_lg")
"Maria Jana ama John Smith e Maria é amada por Joaquim de Souza" |>
  spacyr::spacy_parse(dependency = T) |>
  group_ppn()
}
