% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{group_ppn}
\alias{group_ppn}
\title{Group a POS of proper names}
\usage{
group_ppn(DF)
}
\arguments{
\item{DF}{A data.frame generated by spacyr::spacy_parse().}
}
\description{
Group a data.frame generated by spacyr::spacy_parse(), if there is a sequence of entity = PERS* / PERSON.
It works similar to spacyr::entity_extract(), but preserves dep_rel column
}
\examples{
# example in Portuguese language
# spacy_finalize() # If spacy was previously initialized with another model.
spacy_initialize(model = "pt_core_news_lg")
"Maria Jana ama John Smith e Maria é amada por Joaquim de Souza" |>
  spacyr::spacy_parse(dependency = T) |>
  group_ppn()
}
