% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\docType{data}
\name{layouts}
\alias{layouts}
\title{layouts available to plot the network}
\format{
An object of class \code{character} of length 15.
}
\usage{
layouts
}
\description{
show the displays available to be used in plot functions.
dh: Davidson and Harels algorithm,
mds:multidimensional scaling of nodes
fr: algorithm of Fruchterman and Reingold. igraph::with_fr()
kk: Kamada and Kawai igraph::with_kk(). Kamada, T. and Kawai, S.: An Algorithm for Drawing General Undirected Graphs. _Information Processing Letters_, 31/1, 7-15, 1989.
drl: force directed algorithm from the DrL toolbox igraph::with_drl()
lgl: algorithm from Large Graph Layout. igraph::with_lgl()
}
\keyword{datasets}
