% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_words_in_graph.R
\name{parsePOS}
\alias{parsePOS}
\title{parse tokenized text into POS}
\usage{
parsePOS(txt, bind = TRUE, only_entities = TRUE)
}
\arguments{
\item{txt}{a list of filtered text}

\item{bind}{if TRUE (default), returns a single dataframe. if FALSE, returns a List object}

\item{entities}{if TRUE (default), returns only the entities. if FALSE, returns consolidated entities}
}
\description{
from a list of filtered text (from the function filter_by_query), tags the POS/ parse with spacyr.
It also renumber the doc_id and sentence_id
}
\examples{
txt_wiki |> filter_by_query("Police") |> parsePOS()
txt_wiki |> filter_by_query("Police") |> parsePOS(only_entities = FALSE)
}
