% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph_i.R
\name{plot_graph_i}
\alias{plot_graph_i}
\title{viz graph interactively}
\usage{
plot_graph_i(graph_df, head_n = 20, nodesIdSelection = TRUE, height = "900px")
}
\arguments{
\item{graph_df}{a dataframe with the graph data}

\item{nodesIdSelection}{a boolean value to enable node selection. Default: TRUE.}
}
\description{
Visualize graphs interactively (package visNetwork). The columns must be named:
"from", "label", "to" and "value" (frequency of the triplet) OR the 1st, 2nd
and 3rd columns will be taken as such.
}
\examples{
x <- txt_wiki[2:44] |>
  filter_by_query("Brian") |>
  parsePOS()
g <- get_cooc_entities(x)
g$edges |>
  dplyr::rename(from = n1, to = n2) |>
  viz_graph()
g$edges |> viz_graph()
}
