% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_graph.R
\name{q_plot}
\alias{q_plot}
\title{quick word graph}
\usage{
q_plot(graph_list, color = "lightblue")
}
\arguments{
\item{color}{color of the edges. Default: "lightblue".}

\item{edge_df}{a dataframe of co-occurrence, extracted with `get_cooc_entities()`}
}
\description{
Plot a network of co-ocurrence of terms. For more options, see `plot_graph()`.
}
\examples{
library(txtnet)

g <- txt_wiki[2:44] |> filter_by_query("Police") |> parsePOS()
g <- get_cooc_entities(g)
q_plot(g)
}
