% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{rename_cols}
\alias{rename_cols}
\title{rename cols from count
rename cols froum count to use with other functions, renaming to "from", "to" and "value" for n (frequency.)'}
\usage{
rename_cols(df)
}
\description{
rename cols from count
rename cols froum count to use with other functions, renaming to "from", "to" and "value" for n (frequency.)'
}
