% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_words_in_graph.R
\name{renum_pos}
\alias{renum_pos}
\title{renumerate doc_id and sentence_id}
\usage{
renum_pos(pos_df)
}
\arguments{
\item{pos_df}{a dataframe generated by spacyr::spacy_parse}
}
\description{
renumerate doc_id and sentence_id
}
