% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_to_range.R
\name{scale_to_range}
\alias{scale_to_range}
\title{Change the vector min and max values proportionally}
\usage{
scale_to_range(x, new_min = 0.15, new_max = 8)
}
\arguments{
\item{x}{vector of values}

\item{new_min}{new minimum value. Default 0.15}

\item{new_max}{new maximum value. Default 8}
}
\description{
Used to change the display of values in plot_graph2, to improve visualization
of min and max values
}
\examples{
1:5 |> scale_to_range(0, 1)
1:5 |> scale_to_range(2, 12)
}
