% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{sort_columns_alphabetically}
\alias{sort_columns_alphabetically}
\title{sort columns n1 and n2 alphabetically}
\usage{
sort_columns_alphabetically(DF)
}
\arguments{
\item{DF}{dataframe with n1 and n2 columns}
}
\description{
sort columns n1 and n2 alphabetically
}
