% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{split_graph}
\alias{split_graph}
\title{split a tidy graph (each line is graph with at least 2 nodes) into two dataframes within a list: one with de nodes and its indexes, a second dataframe with de edges.}
\usage{
split_graph(DF_graph)
}
\description{
split a tidy graph (each line is graph with at least 2 nodes) into two dataframes within a list: one with de nodes and its indexes, a second dataframe with de edges.
}
\examples{
DF <- data.frame(from = c("Amanda", "Bruno", "Carlos", "Daniel"), to = c("Bruno", "Carlos", "Daniel", "Amanda"))

split_graph(DF)
}
