% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_link.R
\name{subs_ppn}
\alias{subs_ppn}
\title{Substitute proper names/entities spaces with underscore in the text.}
\usage{
subs_ppn(text, entities, method = "normal")
}
\arguments{
\item{text}{an input text}

\item{entities}{an input vector, as exported by `extract_entity_rb()`}
}
\description{
given a text and a vector of entities, it substitutes the spaces with underscores, so the entities are identified.
}
\examples{
texto_teste <- "José da Silva e Fulano de Tal foram, bla Maria Silva. E depois disso, bla Joaquim José da Silva Xavier no STF"
ppn <- texto_teste |> extract_entity_rb(connectors("pt"), sw = gen_stopwords("pt"))
texto_teste |> subs_ppn(ppn)
texto_teste |> subs_ppn(ppn, method = "loop")
text <- texto_teste |> subs_ppn(ppn)
text
}
