% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblgraph_to_tbl.R
\name{tblgraph_to_tbl}
\alias{tblgraph_to_tbl}
\title{Transform a tbl_graph into a labeled tibble}
\usage{
tblgraph_to_tbl(tbl_graph)
}
\arguments{
\item{tbl_graph}{tbl_graph object, from the package tidygraph}
}
\description{
Transform a tbl_graph into a labeled tibble
}
\examples{
G <- "Lorem ipsum dolor sit amet, consectetur adipiscing elit." |> cooccur()
G
G <- G |> tidygraph::as_tbl_graph()
G
G |> tblgraph_to_tbl()
}
