% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{data}
\name{text_sample}
\alias{text_sample}
\title{A Text Sample}
\format{
An object of class \code{character} of length 1.
}
\usage{
text_sample
}
\description{
A sample text from text from https://www.bbc.com/news/articles/c7ve36zg0e5o
}
\examples{
data(package = "txtnet")
head(text_sample)
}
\keyword{datasets}
