% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto_functions.R
\name{tokenize_words}
\alias{tokenize_words}
\title{tokenize words}
\usage{
tokenize_words(txt, lower = TRUE, unlist = FALSE)
}
\arguments{
\item{txt}{input text}

\item{lower}{convert words to lowercase. Default TRUE.}
}
\description{
A word tokenizer, designed to keep compounded words, like "Covid-19"" or URLs
}
\examples{
txt <- "Was ice-cream better in Soviet Union or in New York? Bla bla, ble. The package https://quanteda.io/reference/index.html is very Nice! There are also compounded words like New_Jersey. The Covid-19 pandemic is very bad. Some,text,like;csv"
tokenize_words(txt)

txt2 <- c("Was ice-cream better before?", "The Covid-19 brought many problems.")
tokenize_words(txt2)
}
