% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_subst.R
\name{txt_subst}
\alias{txt_subst}
\title{Replace text using rule based proper name}
\usage{
txt_subst(txt, connect = connectors("all"), sw = "", underscore = TRUE)
}
\arguments{
\item{txt}{the input text}

\item{connect}{type of connectors}

\item{sw}{the stopwords}

\item{underscore}{default TRUE}
}
\description{
given a text and the proper name connectors (`connectors()`), it returns the text
with its proper names as one single word, with underscore biding them. For 
example, the text "John Does lives in New York in United States of America" 
becomes "John_Does lives in New_York in United_States_of_America."
}
\examples{
txt_subst("John Does lives in New York in United States of America.")

}
