% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txt_wiki}
\alias{txt_wiki}
\title{A Text Sample from wikipedia}
\format{
A vector of text with 43 elements.
}
\source{
<https://en.wikipedia.org/wiki/Killing_of_Brian_Thompson>
}
\usage{
txt_wiki
}
\description{
A vector sample text from Wikipedia
}
\examples{
data(txt_wiki)
str(txt_wiki)
}
\keyword{datasets}
