% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtnet-package.R
\docType{package}
\name{txtnet-package}
\alias{txtnet}
\alias{txtnet-package}
\title{txtnet: build network of text}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SoaresAlisson/txtnet}
  \item Report bugs at \url{https://github.com/SoaresAlisson/txtnet/issues}
}

}
\author{
\strong{Maintainer}: Alisson Soares \email{soares.alisson@yahoo.com.br} (\href{https://orcid.org/0009-0004-5497-5572}{ORCID})

}
\keyword{internal}
