% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{collapse_adp}
\alias{collapse_adp}
\title{Collapse ADP}
\usage{
collapse_adp(pos_df)
}
\arguments{
\item{pos_df}{the POS dataframe generated by spacy_parse}
}
\description{
Collapse upos ADP, if the previous and the next pos are NOUN or PROPN. It is used in the pipeline to extract proper names like "United States of America"
}
\examples{
“The United States of America has issued a new warning to the Labor Court of the State of Mato Grosso do Sul, by way of compensation.” |>
 spacyr::spacy_parse(dependency = T) |>  conflate_adp()

# in Portuguese:
spacy_finalize()
spacyr::spacy_initialize(model = "pt_core_news_lg")
"Os Estados Unidos da América emitiram um novo alerta para a Justiça do Trabalho do Estado do Mato Grosso do Sul, a título de indenização." |>
  spacyr::spacy_parse(dependency = T) |> conflate_adp()
}
