% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_triplets.R
\name{extract_triplets}
\alias{extract_triplets}
\title{from semgram output, join passive and active voices}
\usage{
extract_triplets(semgram)
}
\arguments{
\item{semgram}{a object from spacyr::spacy_parse() |> semgram::extract_motifs()}
}
\description{
from semgram output, join passive and active voices
}
