% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_words_in_graph.R
\name{filter_by_query}
\alias{filter_by_query}
\title{tokenize and filter text by query}
\usage{
filter_by_query(
  txt,
  query,
  i_c = TRUE,
  by_sentence = TRUE,
  unlist = FALSE,
  msg = TRUE
)
}
\arguments{
\item{txt}{vector of texts}

\item{query}{query to filter the text}

\item{by_sentence}{tokenize by sentence (default TRUE). If FALSE, tokenize by paragraph}

\item{unlist}{if TRUE (default FALSE), returns a vector instead of a list object.}

\item{msg}{if TRUE (default TRUE), returns a message if the query is not found.}

\item{ic}{ignore case (default TRUE)}
}
\description{
from vector of texts, tokenize by sentence or paragraph and returns a filtered list
}
\examples{
# loading data
data(package = "txtnet")
# sample
txt_wiki[2:3]
txt_wiki |> filter_by_query("York")
txt_wiki |> filter_by_query("Police")
}
