% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_words_in_graph.R
\name{filter_by_query}
\alias{filter_by_query}
\title{tokenize and filter text by query}
\usage{
filter_by_query(txt, query, ic = TRUE, by_sentence = TRUE, unlist = FALSE)
}
\arguments{
\item{txt}{vector of texts}

\item{query}{query to filter the text}

\item{ic}{ignore case (default TRUE)}

\item{by_sentence}{tokenize by sentence (default TRUE). If FALSE, tokenize by paragraph}

\item{unlist}{if TRUE (default FALSE), returns a vector instead of a list object.}
}
\description{
from vector of texts, tokenize by sentence or paragraph and returns a filtered list
}
\examples{
# loading data
data(package="txtnet")
# sample
txt_wiki[2:3]
txt_wiki |> filter_by_query("York")
txt_wiki |> filter_by_query("Police")
}
