% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_operations.R
\docType{data}
\name{g}
\alias{g}
\title{a sample graph dataframe (tbl-graph)}
\format{
An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 6.
}
\usage{
g
}
\description{
a sample graph dataframe (tbl-graph)
}
\keyword{datasets}
