% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_words_in_graph.R
\name{graph_from_cooccurrence}
\alias{graph_from_cooccurrence}
\title{Graph from word co-occurrence}
\usage{
graph_from_cooccurrence(df_cooccurrence, strip_rgx = "^the_", freq = TRUE)
}
\arguments{
\item{df_cooccurrence}{a dataframe generated by get_pairs()}

\item{strip_rgx}{regex pattern to strip in the node text. Default: "^the_".
To erase nothing, use "".}

\item{freq}{if TRUE (default), returns a dataframe with frequency. If FALSE,
returns only the pairs without aggregation.}
}
\description{
given a dataframe with a column (list type) with POS, returns a list with three
elements: 1) "graphs": a tibble with the frequency of graphs;
2) "isolated_nodes": a tibble with the isolated nodes, i.e. nodes without
any connection.
3) "nodes": a tibble with the individual frequency of each node. If list
element has only one element, it is removed.
}
\examples{
pos <- txt_wiki |>
  filter_by_query("Police") |>
  parsePOS()

entities_by_txt <- pos |>
  dplyr::group_by(doc_id) |>
  dplyr::summarise(entities = list(unique(entity)))

graph_from_cooccurrence(entities_by_txt)
}
