% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_words_in_graph.R
\name{graph_from_coocurrence_list}
\alias{graph_from_coocurrence_list}
\title{graph from co-ocurrence list}
\usage{
graph_from_coocurrence_list(coocurrence_list, strip_rgx = "^the_", freq = TRUE)
}
\arguments{
\item{coocurrence_list}{a dataframe generated by get_pairs()}

\item{strip_rgx}{regex to strip. Default: "^the_". To erase nothing, use "".}

\item{freq}{if TRUE (default), returns a dataframe with frequency. If FALSE, returns only the pairs}
}
\description{
given a listed dataframe of co-ocurrence, returns a dataframe with frequency.
If list element has only one element, it is removed.
}
\examples{
pos <- txt_wiki |>
  filter_by_query("Police") |>
  parsePOS()
entities_by_txt <- pos |>
  dplyr::group_by(doc_id) |>
  dplyr::summarise(entities = list(unique(entity)))
graph_from_coocurrence_list(entities_by_txt)
}
