% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_subs.R
\name{graph_subs}
\alias{graph_subs}
\title{graph substitution}
\usage{
graph_subs(DF, df_subs = NA, rm_symbols = "[^\\\\w\\\\s\\\\_\\\\&]")
}
\arguments{
\item{DF}{a graph data frame}

\item{df_subs}{a data frame with the substitutions to be made, where the
first column is what to substitute and column 2 is what to substitute with.
It uses regular expressions. It can be left empty to use only the rm_symbols
parameter}

\item{rm_symbols}{remove non-word characters (symbols), preserving
letter, numbers and the underscore (to make it clear the words boundaries).
Default regex pattern: "[^\\w\\s\\_\\&]". Is possible to use other custom
regex patterns. If no symbol substitution is desired, leave it empty..}
}
\description{
Substitute node columns (both columns 1 and 2) of a graph data frame with a
dictionary of substitutions.
}
\examples{
# a test tibble
test_graph <- tibble::tibble(
  n1 = c("A", "B", "A", "C", "B", "Ab", "A", "D"),
  n2 = c("B", "Ab", "B", "D", "C", "A", "C", "D")
)

# dataframe with substitutions
DF_substitution <- tibble::tribble(
  ~col1, ~col2,
  "B", "blah",
  "C", "Capybara"
)

# Doing the substitutions
graph_subs(test_graph, DF_substitution)
}
