% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_graph.R
\name{net_wordcloud}
\alias{net_wordcloud}
\title{plot net wordcloud}
\usage{
net_wordcloud(text, df, head_n = 30, color = "lightblue")
}
\arguments{
\item{text}{the original text used to extract the graph. It is necessary to
calculate the individual frequency of the words.}

\item{df}{a dataframe of co-occurrence, extracted with `extract_graph()` and
`count(n1, n2)`}

\item{head_n}{number of nodes to show - the more frequent}
}
\description{
plot a network of co-occurrence of terms, as returned by extract_graph and
then by dplyr::count(). The size of words and compound words means the
individual frequency of each word/compound word. The thickness of the links
indicates how often the pair occur together.
}
\examples{

# stopwords:
my_sw <- c(stopwords::stopwords(language = "en", source = "snowball", simplify = TRUE), "lol")

txt_wiki |> # text available in the package
  # because it is a vector, let's collapse it into a single element:
  paste(collapse = " ") |>
  extract_graph(sw = my_sw) |>
  dplyr::count(n1, n2, sort = TRUE) |> # counting the graphs
  net_wordcloud(txt_wiki, df = _) # plotting
}
