% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\title{plot weighted graph}
\usage{
plot_graph(
  df,
  head_n = 30,
  edge_color = "lightblue",
  edge_alpha = 0.5,
  node_alpha = 0.5,
  text_color = "black",
  text_size = 1
)
}
\arguments{
\item{df}{a dataframe of co-occurrence, extracted with `extract_graph()` and
`count(n1, n2)`}

\item{head_n}{number of nodes to show - the more frequent}

\item{edge_color}{color of the edges}

\item{edge_alpha}{transparency of the edges. Values between 0 and 1.}

\item{node_alpha}{transparency of the nodes}

\item{text_color}{color of the text in nodes}

\item{text_size}{font size of the nodes}

\item{text}{an input text}

\item{scale_graph}{name of a function to normalize the result. Sometime, the range
of numbers are so wide that the graph becomes unreadable. Applying a function
to normalize the result can improve the readability, for example using}
}
\description{
From a dataframe of graph frequence, plot a graph plot a graph of
co-occurrence of terms, as returned by get_cooc_entities, extract_graph
}
\examples{
# plot_graph(df = graph_count, head_n = 50, scale_graph = "log2")
}
