% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph2.R
\name{plot_graph2}
\alias{plot_graph2}
\title{weighted word graph viz}
\usage{
plot_graph2(
  text,
  df,
  head_n = 30,
  edge_color = "lightblue",
  edge_alpha = 0.5,
  node_alpha = 0.5,
  text_color = "black",
  text_size = 1,
  scale_graph = "scale_values"
)
}
\arguments{
\item{text}{an input text}

\item{df}{a dataframe of co-occurrence, extracted with `extract_graph()` and
`count(n1, n2)`}

\item{head_n}{number of nodes to show - the more frequent}

\item{edge_color}{color of the edges}

\item{edge_alpha}{transparency of the edges. Values between 0 and 1.}

\item{node_alpha}{transparency of the nodes}

\item{text_color}{color of the text in nodes}

\item{text_size}{font size of the nodes}

\item{scale_graph}{name of a function to normalize the result. Sometime, the range
of numbers are so wide that the graph becomes unreadable. Applying a function
to normalize the result can improve the readability, for example using
`scale_graph = "log2"`, or `"log10"`}
}
\description{
Plot a network of co-occurrence of terms. Because word frequencies can vary
significantly, differences in text size can be substantial. Therefore,
instead of adjusting text size, we vary the dot/node size, ensuring the
text remains consistently sized and maintains readability. It is also
possible to normalize the result with log.
}
\details{
plot a graph of co-occurrence of terms, as returned by extract_graph
}
\examples{
# plot_graph(txt, df = graph_count, head_n = 50, scale_graph = "log2")
}
