% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_graph.R
\name{plot_graph2}
\alias{plot_graph2}
\title{weighted graph viz
plot a network of co-ocurrence of terms}
\usage{
plot_graph2(text, df, head_n = 30, color = "lightblue", scale = "scale_values")
}
\arguments{
\item{text}{an input text}

\item{df}{a dataframe of co-occurrence, extracted with `extract_graph()` and `count(n1, n2)`}

\item{head_n}{number of nodes to show - the more frequent}

\item{scale}{name of a function to normalize the result. Sometime, the range of numbers are so wide that the graph becomes unreadable. Applying a function to normalize the result can improve the readability, for example using `scale = "log2"`, or "log10"}
}
\description{
plot a graph of co-occurrence of terms, as returned by extract_graph
}
\examples{
# plot_graph(txt, df= graph_count, head_n = 50, scale = "log2")
}
